Modify ulr_metrics call to explicitely define metrics
An API change by SEOmoz meant this method was no longer returning
any data, see https://github.com/adamtaylor/WWW-SEOmoz/issues/3.

We now have to explicitely request the data we want back so we've
added some default values to the request. Should be extended in
the future to allow the user to request the data they want.